refactor: rename Fetcher/index[fetcher].js to Fetcher/index.Fetcher.js
